:recycle: Refactor MemberSelectModal
