+++ modals/ReservationModal/Page/MemberSelect/ButtonContainer/index.tsx
@@ -1,0 +1,12 @@
+import { NextButton } from './NextButton';
+import PrevButton from './PrevButton';
+import * as S from './style';
+
+export default function ButtionContainer() {
+    return (
+      <S.ButtonContainer>
+       <PrevButton />
+        <NextButton />
+      </S.ButtonContainer>
+    );
+}
