+++ modals/ReservationModal/Page/MemberSelect/ButtonContainer/style.ts
@@ -1,0 +1,9 @@
+import styled from '@emotion/styled'
+
+export const ButtonContainer = styled.div`
+  position: absolute;
+  bottom: 0;
+  width: 100%;
+  display: flex;
+  justify-content: space-between;
+`
